# backend/openapi.seoul-walking-path-mvp.yaml
openapi: 3.0.3
info:
  title: 서울걷길 관광지 수집형 MVP API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /cards/catalog:
    get:
      summary: 카드 도감 목록
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 카드 도감 반환
  /cards/my:
    get:
      summary: 사용자 수집 카드 목록
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      responses:
        '200':
          description: 수집 카드 목록 반환
...